Rename kernel_types.hpp to kernel.hpp
