+++ include/BURST/kernel.hpp
